refactor: create sub widgets
