+++ lib/app/features/chat/messages/views/components/message_reactions/message_reactions.dart
@@ -1,11 +1,15 @@
 // SPDX-License-Identifier: ice License 1.0
 
 import 'package:flutter/material.dart';
+import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:ion/app/components/avatar/avatar.dart';
 import 'package:ion/app/extensions/extensions.dart';
+import 'package:ion/app/features/auth/providers/auth_provider.dart';
 import 'package:ion/app/features/chat/model/message_reaction_group.dart';
 import 'package:ion/app/features/user/providers/user_metadata_provider.dart';
+
+part 'components/message_reaction_chip.dart';
 
 class MessageReactions extends StatelessWidget {
   const MessageReactions({required this.reactions, super.key});
@@ -30,61 +34,6 @@ class MessageReactions extends StatelessWidget {
               ),
             )
             .toList(),
-      ),
-    );
-  }
-}
-
-class _MessageReactionChip extends ConsumerWidget {
-  const _MessageReactionChip({
-    required this.emoji,
-    required this.pubkeys,
-  });
-
-  final String emoji;
-  final List<String> pubkeys;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final userPictures = pubkeys
-        .map((pubkey) => ref.watch(userMetadataProvider(pubkey)).valueOrNull?.data.picture)
-        .toList();
-
-    return Container(
-      padding: EdgeInsets.symmetric(vertical: 4.0.s, horizontal: 6.0.s),
-      decoration: BoxDecoration(
-        color: context.theme.appColors.primaryBackground,
-        borderRadius: BorderRadius.circular(10.0.s),
-      ),
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(emoji, style: context.theme.appTextThemes.title.copyWith(height: 1)),
-          Stack(
-            children: [
-              //TODO: show only 3 reactions if there are more than 3 reactions when figma is ready
-              for (int i = userPictures.length - 1; i >= 0; i--)
-                Padding(
-                  padding: EdgeInsets.only(left: i * 8.0.s),
-                  child: DecoratedBox(
-                    decoration: BoxDecoration(
-                      color: context.theme.appColors.onPrimaryAccent,
-                      borderRadius: BorderRadius.circular(10),
-                    ),
-                    child: Padding(
-                      padding: EdgeInsets.all(1.0.s),
-                      child: Avatar(
-                        size: 16.0.s,
-                        borderRadius: BorderRadius.circular(10),
-                        imageUrl: userPictures[i],
-                      ),
-                    ),
-                  ),
-                ),
-            ],
-          ),
-        ],
       ),
     );
   }
